0.0.024
offline button spawn updated.
changed how it works to now be loaded by another object at the start.
script now keeps track of information such as type and position in a
struct.
GameObject buttons are deleted through scene changes (because they are
children of other objects and thus can't be saved) but are recreated
when the scene is loaded again by the new game building controller
(calling the reloadButtons function in the other script).