0.0.25d
Fixed online mode so it checks Google Places API and spawns clickable
areas accordingly.  Online mode is now a separate scene.  Currently
correct spawns on the y axis with respect to Latitude.  Still working on
it not spawning in the same spot on the x axis with respect to
longitude.  Last thing to add is the re-spawn when going back from area
exploration and the refresh when player location has changed.